refactor(export): extract the georeferencing phase into step-georeferencing.ts (#2475 step 2a)
Moves `export()`'s IfcProjectedCRS / IfcMapConversion phase, its six helpers
(`resolveMapUnitReference`, `normalizeMapUnitName`, `findLengthUnitReference`,
`reportMapConversionRefused`, `reportMapConversionRefusedNoCrs`,
`findPreferredGeometricRepresentationContextId`) and the two
`MAP_CONVERSION_WITHOUT_*_WARNING` constants out of `step-exporter.ts`, which
drops 2672 -> 2392 lines. `step-exporter.ts` keeps a one-line call site.

The phase is not a pure function: it allocates express ids from the exporter's
`nextExpressId` and reads the store through its `EntityExtractor`. Those are
injected as a `GeorefContext` callback object, following
`type-owned-psets.ts:resolveTypeOwnedPsetIds` and its `nameOf` callback, rather
than hoisted onto `ExportPass` — the counter is incremented at six further
sites in the property/quantity-set generators, so moving it would change what
it computes, not merely where it is named. `applyGeoreferencingMutations`
mutates the pass in place because two consumers read its output at a distance:
the `deltaOnly` empty-export early return and the assembly loop.

The 15 tests that cover the phase move with it into
`step-georeferencing.test.ts`, keeping their original `describe` names so the
suite reports identical full test names. `rejects georeferencing edits for
IFC2X3 export` stays behind (it throws in setup, before this phase runs), as
does `reuses the project length unit when exporting property units` (it reaches
`findLengthUnitReference` through `findUnitId` on the property-set path).

Verified per the standing per-site rule: all 43 guards and ledger call sites in
the moved region were mutated individually before and after the move, and every
one killed a byte-identical set of test NAMES. Suite 738 tests / 708 passed /
30 skipped before and after, with identical full-name sets. No behaviour
change, so no changeset.
