+++ packages/export/src/step-exporter.test.ts
@@ -72,22 +72,6 @@ function buildMockDataStore(
   } as unknown as IfcDataStore;
 }
 
-/** A minimal already-georeferenced model: `#40` CRS, `#41` MapConversion. */
-function buildGeoreferencedMockDataStore(): IfcDataStore {
-  return buildMockDataStore([
-    [1, 'IFCPROJECT', "#1=IFCPROJECT('g',$,'Project',$,$,$,$,(#20),#30);"],
-    [2, 'IFCSIUNIT', '#2=IFCSIUNIT(*,.LENGTHUNIT.,$,.METRE.);'],
-    [20, 'IFCGEOMETRICREPRESENTATIONCONTEXT', "#20=IFCGEOMETRICREPRESENTATIONCONTEXT($,'Model',3,1.E-05,#21,$);"],
-    [21, 'IFCAXIS2PLACEMENT3D', '#21=IFCAXIS2PLACEMENT3D(#22,#23,#24);'],
-    [22, 'IFCCARTESIANPOINT', '#22=IFCCARTESIANPOINT((0.,0.,0.));'],
-    [23, 'IFCDIRECTION', '#23=IFCDIRECTION((0.,0.,1.));'],
-    [24, 'IFCDIRECTION', '#24=IFCDIRECTION((1.,0.,0.));'],
-    [30, 'IFCUNITASSIGNMENT', '#30=IFCUNITASSIGNMENT((#2));'],
-    [40, 'IFCPROJECTEDCRS', "#40=IFCPROJECTEDCRS('EPSG:2056',$,'CH1903+',$,$,$,#2);"],
-    [41, 'IFCMAPCONVERSION', '#41=IFCMAPCONVERSION(#20,#40,1.,2.,3.,1.,0.,1.);'],
-  ]);
-}
-
 /** Count `#N` references in the output that have no `#N=` definition. */
 function findDanglingRefs(content: string): number[] {
   const defined = new Set<number>();
@@ -242,291 +226,6 @@ describe('StepExporter', () => {
     expect(decode(result.content)).not.toContain("#114=IFCPROPERTYSET('3wkd_mjInDCfOthy7w_A6V'");
     expect(decode(result.content)).not.toMatch(/IFCRELDEFINESBYPROPERTIES\([^;]*\(#67\),#/);
     expect(decode(result.content)).toMatch(/#67=IFCWALLTYPE\([^;]*\(#72,#\d+\)[^;]*\);/);
-  });
-
-  it('creates IfcProjectedCRS and IfcMapConversion from scratch when georeferencing is added', async () => {
-    const parser = new IfcParser();
-    const store = await parser.parseColumnar(new TextEncoder().encode(SIMPLE_TYPE_INHERITANCE_IFC).buffer);
-    const exporter = new StepExporter(store);
-
-    const result = exporter.export({
-      schema: 'IFC4',
-      applyMutations: true,
-      georefMutations: {
-        projectedCRS: {
-          name: 'EPSG:2056',
-          description: 'CH1903+ / LV95',
-          geodeticDatum: 'CH1903+',
-          mapProjection: 'Swiss Oblique Mercator 1995',
-          mapUnit: 'METRE',
-        },
-        mapConversion: {
-          eastings: 2600000,
-          northings: 1200000,
-          orthogonalHeight: 500,
-          xAxisAbscissa: 0,
-          xAxisOrdinate: 1,
-          scale: 1,
-        },
-      },
-    });
-
-    const content = decode(result.content);
-    expect(content).toContain("IFCPROJECTEDCRS('EPSG:2056','CH1903+ / LV95','CH1903+',$,'Swiss Oblique Mercator 1995',$,#");
-    expect(content).toMatch(/IFCMAPCONVERSION\(#14,#\d+,2600000\.,1200000\.,500\.,0\.,1\.,1\.\);/);
-    expect(content).toContain('IFCSIUNIT(*,.LENGTHUNIT.,$,.METRE.)');
-    // A surviving context means nothing was refused — no spurious warning.
-    expect(result.stats.warnings).toEqual([]);
-  });
-
-  it('prefers the 3D model representation context when creating IfcMapConversion', () => {
-    const dataStore = buildMockDataStore([
-      [1, 'IFCPROJECT', "#1=IFCPROJECT('g',$,'Project',$,$,$,$,(#10,#20),#30);"],
-      [10, 'IFCGEOMETRICREPRESENTATIONCONTEXT', "#10=IFCGEOMETRICREPRESENTATIONCONTEXT($,'Plan',2,1.E-05,#11,$);"],
-      [11, 'IFCAXIS2PLACEMENT2D', "#11=IFCAXIS2PLACEMENT2D(#12,#13);"],
-      [12, 'IFCCARTESIANPOINT', '#12=IFCCARTESIANPOINT((0.,0.));'],
-      [13, 'IFCDIRECTION', '#13=IFCDIRECTION((1.,0.));'],
-      [20, 'IFCGEOMETRICREPRESENTATIONCONTEXT', "#20=IFCGEOMETRICREPRESENTATIONCONTEXT($,'Model',3,1.E-05,#21,$);"],
-      [21, 'IFCAXIS2PLACEMENT3D', "#21=IFCAXIS2PLACEMENT3D(#22,#23,#24);"],
-      [22, 'IFCCARTESIANPOINT', '#22=IFCCARTESIANPOINT((0.,0.,0.));'],
-      [23, 'IFCDIRECTION', '#23=IFCDIRECTION((0.,0.,1.));'],
-      [24, 'IFCDIRECTION', '#24=IFCDIRECTION((1.,0.,0.));'],
-      [30, 'IFCUNITASSIGNMENT', '#30=IFCUNITASSIGNMENT(());'],
-    ]);
-
-    const exporter = new StepExporter(dataStore);
-    const result = exporter.export({
-      schema: 'IFC4',
-      applyMutations: true,
-      georefMutations: {
-        projectedCRS: { name: 'EPSG:2056', mapUnit: 'METRE' },
-        mapConversion: { eastings: 2600000, northings: 1200000, orthogonalHeight: 500, xAxisAbscissa: 1, xAxisOrdinate: 0, scale: 1 },
-      },
-    });
-
-    expect(decode(result.content)).toMatch(/IFCMAPCONVERSION\(#20,#\d+,2600000\.,1200000\.,500\.,1\.,0\.,1\.\);/);
-    // Happy path: the conversion was written, so nothing was refused.
-    expect(result.stats.warnings).toEqual([]);
-  });
-
-  describe('map conversion that cannot be written (#2067)', () => {
-    // No IfcGeometricRepresentationContext anywhere, so there is no id the new
-    // IfcMapConversion could use as SourceCRS. Skipping it is correct — writing
-    // it would dangle — but the file that comes back is identical to one where
-    // no map conversion was ever requested, so the refusal has to be returned.
-    const NO_CONTEXT_ENTRIES: Array<[number, string, string]> = [
-      [1, 'IFCPROJECT', "#1=IFCPROJECT('g',$,'Project',$,$,$,$,$,#30);"],
-      [2, 'IFCSIUNIT', '#2=IFCSIUNIT(*,.LENGTHUNIT.,$,.METRE.);'],
-      [30, 'IFCUNITASSIGNMENT', '#30=IFCUNITASSIGNMENT((#2));'],
-    ];
-
-    it('reports the refusal when creating both the CRS and the conversion', () => {
-      const dataStore = buildMockDataStore(NO_CONTEXT_ENTRIES);
-
-      const result = new StepExporter(dataStore).export({
-        schema: 'IFC4',
-        applyMutations: true,
-        georefMutations: {
-          projectedCRS: { name: 'EPSG:1234', mapUnit: 'METRE' },
-          mapConversion: { eastings: 2600000, northings: 1200000, orthogonalHeight: 500, xAxisAbscissa: 1, xAxisOrdinate: 0, scale: 1 },
-        },
-      });
-
-      const content = decode(result.content);
-      // A map conversion really was requested and the CRS really was written —
-      // the absent IFCMAPCONVERSION is the thing under test, not a no-op export.
-      expect(content).toContain("IFCPROJECTEDCRS('EPSG:1234'");
-      expect(content).not.toContain('IFCMAPCONVERSION');
-      expect(findDanglingRefs(content)).toEqual([]);
-      expect(result.stats.warnings).toHaveLength(1);
-      // `SourceCRS` is the substring that discriminates this refusal from the
-      // TargetCRS one: `toContain('IfcMapConversion')` alone is satisfied by
-      // either message, so it cannot tell the two refusals apart (#2105
-      // review). Full equality pins the explanation, not merely that
-      // something was reported.
-      expect(result.stats.warnings[0]).toContain('SourceCRS');
-      expect(result.stats.warnings[0]).toBe(
-        'Cannot create IfcMapConversion: no IfcGeometricRepresentationContext is available to reference as SourceCRS. The IfcProjectedCRS is unaffected.',
-      );
-    });
-
-    it('reports the refusal when the CRS already exists and only the conversion is new', () => {
-      const dataStore = buildMockDataStore([
-        ...NO_CONTEXT_ENTRIES,
-        [40, 'IFCPROJECTEDCRS', "#40=IFCPROJECTEDCRS('EPSG:1234',$,$,$,$,$,#2);"],
-      ]);
-
-      const result = new StepExporter(dataStore).export({
-        schema: 'IFC4',
-        applyMutations: true,
-        georefMutations: {
-          mapConversion: { eastings: 2600000, northings: 1200000, orthogonalHeight: 500 },
-        },
-      });
-
-      const content = decode(result.content);
-      expect(content).toContain("#40=IFCPROJECTEDCRS('EPSG:1234'");
-      expect(content).not.toContain('IFCMAPCONVERSION');
-      expect(result.stats.warnings).toHaveLength(1);
-      // On this path the export writes no IfcProjectedCRS at all — it
-      // references the pre-existing #40. The message must not claim a CRS
-      // was written, only that the conversion could not be (#2105 review).
-      // Full equality pins the explanation to the same strength as the other
-      // two refusal tests in this describe block (#2105 review round two).
-      expect(result.stats.warnings[0]).not.toContain('was written');
-      expect(result.stats.warnings[0]).toBe(
-        'Cannot create IfcMapConversion: no IfcGeometricRepresentationContext is available to reference as SourceCRS. The IfcProjectedCRS is unaffected.',
-      );
-    });
-
-    it('reports the refusal even when the delta export has nothing else to write', () => {
-      const dataStore = buildMockDataStore([
-        ...NO_CONTEXT_ENTRIES,
-        [40, 'IFCPROJECTEDCRS', "#40=IFCPROJECTEDCRS('EPSG:1234',$,$,$,$,$,#2);"],
-      ]);
-
-      const result = new StepExporter(dataStore).export({
-        schema: 'IFC4',
-        applyMutations: true,
-        deltaOnly: true,
-        georefMutations: {
-          mapConversion: { eastings: 2600000, northings: 1200000, orthogonalHeight: 500 },
-        },
-      });
-
-      expect(decode(result.content)).not.toContain('IFCMAPCONVERSION');
-      expect(result.stats.warnings).toHaveLength(1);
-    });
-
-    it('reports nothing when no georeferencing was requested', () => {
-      const dataStore = buildMockDataStore(NO_CONTEXT_ENTRIES);
-
-      const result = new StepExporter(dataStore).export({ schema: 'IFC4', applyMutations: true });
-
-      expect(result.stats.warnings).toEqual([]);
-    });
-
-    it('reports the refusal when a map conversion is requested with no projectedCRS and none in the file', () => {
-      // Neither IFCPROJECTEDCRS nor IFCMAPCONVERSION exists, and the caller
-      // did not ask for a projectedCRS either — both CREATE branches skip,
-      // so nothing is attempted and (before the fix) nothing was refused.
-      //
-      // The fixture CARRIES a context on purpose (review of #2105). Under
-      // NO_CONTEXT_ENTRIES both refusal conditions are true at once, so the
-      // context-specific message reads as acceptable here and the test cannot
-      // tell the two refusals apart — it passed when the branch emitted the
-      // wrong explanation. With a usable context present, the missing CRS is
-      // the only thing that can produce a refusal.
-      const dataStore = buildMockDataStore([
-        ...NO_CONTEXT_ENTRIES,
-        [50, 'IFCGEOMETRICREPRESENTATIONCONTEXT', "#50=IFCGEOMETRICREPRESENTATIONCONTEXT($,'Model',3,1.E-05,$,$);"],
-      ]);
-
-      const result = new StepExporter(dataStore).export({
-        schema: 'IFC4',
-        applyMutations: true,
-        georefMutations: {
-          mapConversion: { eastings: 2600000, northings: 1200000, orthogonalHeight: 500 },
-        },
-      });
-
-      const content = decode(result.content);
-      expect(content).not.toContain('IFCMAPCONVERSION');
-      expect(content).not.toContain('IFCPROJECTEDCRS');
-      expect(result.stats.warnings).toHaveLength(1);
-      // `TargetCRS` is the one substring that discriminates the two messages:
-      // the context message also mentions both IfcMapConversion and
-      // IfcProjectedCRS ("...The IfcProjectedCRS is unaffected"), so asserting
-      // on those alone passes for either. Equality pins the explanation, not
-      // merely that something was reported.
-      expect(result.stats.warnings[0]).toContain('TargetCRS');
-      expect(result.stats.warnings[0]).toBe(
-        'Cannot create IfcMapConversion: no IfcProjectedCRS was requested and none exists in the file to reference as TargetCRS. Nothing was written.',
-      );
-    });
-  });
-
-  it('recreates georeferencing the session deleted instead of editing the tombstone', () => {
-    const dataStore = buildGeoreferencedMockDataStore();
-    const mutationView = new LiveMutablePropertyView(null, 'model-georef');
-    mutationView.deleteEntity(40);
-    mutationView.deleteEntity(41);
-
-    const exporter = new StepExporter(dataStore, mutationView);
-    const result = exporter.export({
-      schema: 'IFC4',
-      applyMutations: true,
-      georefMutations: {
-        projectedCRS: { name: 'EPSG:1234', mapUnit: 'METRE' },
-        mapConversion: { eastings: 10, northings: 20, orthogonalHeight: 30, xAxisAbscissa: 1, xAxisOrdinate: 0, scale: 1 },
-      },
-    });
-
-    const content = decode(result.content);
-    // The tombstoned #40/#41 are never written, so the edit has to land on a
-    // NEW pair — otherwise the georeferencing disappears from the file.
-    expect(content).not.toContain('#40=');
-    expect(content).not.toContain('#41=');
-    expect(content).toContain("IFCPROJECTEDCRS('EPSG:1234'");
-    expect(content).toMatch(/IFCMAPCONVERSION\(#20,#\d+,10\.,20\.,30\.,1\.,0\.,1\.\);/);
-    expect(findDanglingRefs(content)).toEqual([]);
-  });
-
-  it('recreates a deleted IfcMapConversion while keeping the surviving IfcProjectedCRS', () => {
-    const dataStore = buildGeoreferencedMockDataStore();
-    const mutationView = new LiveMutablePropertyView(null, 'model-georef');
-    mutationView.deleteEntity(41);
-
-    const exporter = new StepExporter(dataStore, mutationView);
-    const result = exporter.export({
-      schema: 'IFC4',
-      applyMutations: true,
-      georefMutations: {
-        mapConversion: { eastings: 10, northings: 20, orthogonalHeight: 30, xAxisAbscissa: 1, xAxisOrdinate: 0, scale: 1 },
-      },
-    });
-
-    const content = decode(result.content);
-    expect(content).not.toContain('#41=');
-    expect(content).toMatch(/IFCMAPCONVERSION\(#20,#40,10\.,20\.,30\.,1\.,0\.,1\.\);/);
-    expect(findDanglingRefs(content)).toEqual([]);
-  });
-
-  it('never points new georeferencing at a deleted context or length unit', () => {
-    const dataStore = buildMockDataStore([
-      [1, 'IFCPROJECT', "#1=IFCPROJECT('g',$,'Project',$,$,$,$,(#20,#25),#30);"],
-      [2, 'IFCSIUNIT', '#2=IFCSIUNIT(*,.LENGTHUNIT.,$,.METRE.);'],
-      [20, 'IFCGEOMETRICREPRESENTATIONCONTEXT', "#20=IFCGEOMETRICREPRESENTATIONCONTEXT($,'Model',3,1.E-05,#21,$);"],
-      [21, 'IFCAXIS2PLACEMENT3D', '#21=IFCAXIS2PLACEMENT3D(#22,#23,#24);'],
-      [22, 'IFCCARTESIANPOINT', '#22=IFCCARTESIANPOINT((0.,0.,0.));'],
-      [23, 'IFCDIRECTION', '#23=IFCDIRECTION((0.,0.,1.));'],
-      [24, 'IFCDIRECTION', '#24=IFCDIRECTION((1.,0.,0.));'],
-      [25, 'IFCGEOMETRICREPRESENTATIONCONTEXT', "#25=IFCGEOMETRICREPRESENTATIONCONTEXT($,'Model',3,1.E-05,#21,$);"],
-      [30, 'IFCUNITASSIGNMENT', '#30=IFCUNITASSIGNMENT((#2));'],
-    ]);
-    const mutationView = new LiveMutablePropertyView(null, 'model-georef');
-    mutationView.deleteEntity(20);
-    mutationView.deleteEntity(2);
-
-    const exporter = new StepExporter(dataStore, mutationView);
-    const result = exporter.export({
-      schema: 'IFC4',
-      applyMutations: true,
-      georefMutations: {
-        projectedCRS: { name: 'EPSG:1234', mapUnit: 'METRE' },
-        mapConversion: { eastings: 10, northings: 20, orthogonalHeight: 30, xAxisAbscissa: 1, xAxisOrdinate: 0, scale: 1 },
-      },
-    });
-
-    const content = decode(result.content);
-    // SourceCRS falls through to the surviving context, and the map unit is
-    // synthesised rather than reusing the tombstoned #2.
-    expect(content).toMatch(/IFCMAPCONVERSION\(#25,#\d+,10\.,20\.,30\.,1\.,0\.,1\.\);/);
-    const crsLine = content.match(/#\d+=IFCPROJECTEDCRS\([^;]*\);/)?.[0] ?? '';
-    expect(crsLine).toContain("'EPSG:1234'");
-    expect(crsLine).not.toContain('#2)');
-    const newUnitId = Number(crsLine.match(/#(\d+)\);$/)?.[1]);
-    expect(content).toContain(`#${newUnitId}=IFCSIUNIT(*,.LENGTHUNIT.,$,.METRE.);`);
   });
 
   it('rejects georeferencing edits for IFC2X3 export', async () => {
